@@ -18,11 +18,11 @@
         }
     },
     "require": {
-        "php": ">=5.5"
+        "php": ">=7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.27",
-        "squizlabs/php_codesniffer": "^2.7",
-        "php-coveralls/php-coveralls": "^1.0"
+        "php-coveralls/php-coveralls": "^1.0",
+        "phpunit/phpunit": "^6.5",
+        "squizlabs/php_codesniffer": "^3.2"
     }
 }
