@@ -16,8 +16,8 @@ abstract class Value
         OptionalStr::class     => self::TYPE_STRING,
         Number::class          => self::TYPE_NUMBER,
         OptionalNumber::class  => self::TYPE_NUMBER,
-        Object::class          => self::TYPE_OBJECT,
-        OptionalObject::class  => self::TYPE_OBJECT,
+        Obj::class          => self::TYPE_OBJECT,
+        OptionalObj::class  => self::TYPE_OBJECT,
         Arr::class             => self::TYPE_ARRAY,
         OptionalArr::class     => self::TYPE_ARRAY,
         Boolean::class         => self::TYPE_BOOLEAN,
